+++ src/pages/mod.rs
@@ -24,6 +24,7 @@ pub mod dispute_inspector;
 pub mod explorer;
 pub mod fixture_lab;
 pub mod object_inspector;
+pub mod object_page;
 pub mod piteka_environment;
 pub mod tuppira_explorer;
 
@@ -53,6 +54,7 @@ pub use explorer::Explorer;
 pub use fixture_lab::FixtureLab;
 pub use nft_page::{NftCollections, NftDetail, NftGallery};
 pub use object_inspector::ObjectInspector;
+pub use object_page::ObjectPage;
 pub use piteka_environment::PitekaEnvironmentReceipt;
 pub use tuppira_explorer::TuppiraExplorer;
 pub use wallet_page::WalletPage;